requirements.txt
Flask
flask-cors==4.0.0
requests==2.31.0
python-dotenv